feat: Created authorization for only admin of club (starting)

+++ API-PI-Clubes/Controllers/CourtController.cs
@@ -2,6 +2,7 @@
 using API_PI_Clubes.Application.Interfaces.IServices;
 using API_PI_Clubes.Infrastructure.Data;
 using API_PI_Clubes.Model;
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -13,9 +14,11 @@ namespace API_PI_Clubes.Controllers
     public class CourtController : ControllerBase
     {
         private readonly ICourtService _service;
-        public CourtController(ICourtService service)
+        private readonly IAuthorizationService _authorizationService;
+        public CourtController(ICourtService service, IAuthorizationService authorizationService)
         {
             _service = service;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -53,10 +56,19 @@ namespace API_PI_Clubes.Controllers
             var result = await _service.Update(id, dto);
             return Ok(result);
         }
-        [Authorize(Roles = "Admin")]
+        [Authorize(Roles = "Admin,Both")]
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(Guid id)
         {
+            var data = await _service.GetById(id);
+            var clubId = data.ClubId;
+            var isAuthorized = await _authorizationService
+                .AuthorizeAsync(User, clubId, "AdminClubPolicy");
+            if (!isAuthorized.Succeeded)
+            {
+                return Forbid();
+            }
+            
             await _service.Delete(id);
             return NoContent();
         }
